+++ react-components/src/architecture/base/concreteCommands/pointcloud/SetPointColorTypeCommand.ts
@@ -3,9 +3,9 @@
  */
 
 import { PointColorType } from '@cognite/reveal';
-import { BaseOptionCommand } from '../commands/BaseOptionCommand';
-import { RenderTargetCommand } from '../commands/RenderTargetCommand';
-import { type TranslateKey } from '../utilities/TranslateKey';
+import { BaseOptionCommand } from '../../commands/BaseOptionCommand';
+import { RenderTargetCommand } from '../../commands/RenderTargetCommand';
+import { type TranslateKey } from '../../utilities/TranslateKey';
 
 const DEFAULT_OPTIONS: PointColorType[] = [
   PointColorType.Rgb,
